TexGaussian/requirements.txt
argparse
tyro
diffusers
einops
imageio
imageio-ffmpeg
matplotlib
numpy
opencv-python
Pillow
plyfile
scikit-image
scikit-learn
scipy
timm
tokenizers
tqdm
ftfy
kornia
dearpygui
accelerate
gradio
lpips
packaging
pygltflib
rembg[gpu,cli]
rich
safetensors
transformers
trimesh
kiui
xatlas
roma
ocnn
open3d
bpy
blendertoolbox
tensorboard
pymeshlab
argparse